align code with float4_e2m1
