+++ src/core/include/openvino/core/type/float8_e8m0.hpp
@@ -4,7 +4,6 @@
 
 #pragma once
 
-#include <climits>
 #include <cmath>
 #include <iostream>
 #include <limits>
@@ -20,39 +19,12 @@ namespace ov {
  * @brief Class to represent the f8e8m0 type.
  */
 class OPENVINO_API float8_e8m0 {
-    uint8_t f32_to_f8e8m0_bits(const float value) {
-        constexpr uint8_t f32_mantissa_bits{23u};
-        constexpr uint32_t f32_exponent_bits_mask{0x7f800000u};
-        constexpr uint32_t f32_mantissa_bits_mask{0x007fffffu};
-        constexpr uint32_t round_even{0x00400000u};
-
-        const auto input_bits = *reinterpret_cast<const uint32_t*>(&value);
-        const auto input_exponent_bits =
-            static_cast<uint8_t>((input_bits & f32_exponent_bits_mask) >> f32_mantissa_bits);
-
-        if (value <= 0.0) {
-            return 0b00000000;
-        } else if (std::isinf(value) || input_exponent_bits == 0b11111110) {
-            return 0b11111110;
-        } else if (std::isnan(value)) {
-            return 0b11111111;
-        }
-
-        if ((input_bits & f32_mantissa_bits_mask) > round_even) {
-            return input_exponent_bits + 1;
-        } else if ((input_bits & f32_mantissa_bits_mask) == round_even) {
-            return input_exponent_bits + (input_exponent_bits & 0x1);
-        } else {
-            return input_exponent_bits;
-        }
-    }
-
 public:
     float8_e8m0() = default;
-    float8_e8m0(const float value) : m_value{f32_to_f8e8m0_bits(value)} {};
+    float8_e8m0(const float value);
 
     template <typename I>
-    explicit float8_e8m0(I value) : m_value{float8_e8m0{static_cast<float>(value)}.m_value} {}
+    explicit float8_e8m0(I value) : float8_e8m0{static_cast<float>(value)} {}
 
     template <typename T>
     bool operator==(const T& other) const;
@@ -86,30 +58,14 @@ public:
     template <typename T>
     float8_e8m0 operator/=(const T& other);
 
-    operator float() const {
-        constexpr uint8_t float_mantissa_bits{23u};
-        constexpr uint8_t byte_mask{0xffu};
-        union {
-            uint32_t i_val;
-            float f_val;
-        };
-
-        if (to_bits() == 0xff) {
-            return std::numeric_limits<float>::quiet_NaN();
-        } else if (to_bits() == 0x00) {
-            return std::numeric_limits<float>::min() / 2;
-        }
-
-        i_val = static_cast<uint32_t>(m_value & byte_mask) << float_mantissa_bits;
-        return f_val;
-    }
+    operator float() const;
 
     static constexpr float8_e8m0 from_bits(uint8_t bits) {
         return float8_e8m0(bits, true);
     }
-    uint8_t to_bits() const {
-        return m_value;
-    }
+
+    uint8_t to_bits() const;
+
     friend std::ostream& operator<<(std::ostream& out, const float8_e8m0& obj) {
         out << static_cast<float>(obj);
         return out;
@@ -243,22 +199,20 @@ public:
     static constexpr float_denorm_style has_denorm = denorm_absent;
     static constexpr bool has_denorm_loss = false;
 
+    static constexpr ov::float8_e8m0 infinity() noexcept {
+        return ov::float8_e8m0::from_bits(0);  // no infinity
+    }
     static constexpr ov::float8_e8m0 quiet_NaN() noexcept {
         return ov::float8_e8m0::from_bits(0b11111111);
     }
-
+    static constexpr ov::float8_e8m0 signaling_NaN() noexcept {
+        return ov::float8_e8m0::from_bits(0);  // no signaling NaN
+    }
     static constexpr bool is_iec559 = false;
-    static constexpr bool is_bounded = true;
+    static constexpr bool is_bounded = false;
     static constexpr bool is_modulo = false;
     static constexpr bool traps = false;
     static constexpr bool tinyness_before = false;
     static constexpr float_round_style round_style = round_to_nearest;
 };
 }  // namespace std
-
-static_assert(sizeof(ov::float8_e8m0) == 1, "class f8e8m0 must be exactly 1 byte");
-static_assert(std::is_trivially_constructible<ov::float8_e8m0, ov::float8_e8m0>::value, "should be trivially constructible");
-static_assert(std::is_trivially_copyable<ov::float8_e8m0>::value, "must be trivially copyable");
-static_assert(std::is_trivially_destructible<ov::float8_e8m0>::value, "must be trivially destructible");
-static_assert(std::numeric_limits<ov::float8_e8m0>::is_specialized, "numeric_limits must be specialized");
-static_assert(!std::numeric_limits<ov::float8_e8m0>::is_integer, "numeric_limits::is_integer must be false");
